// src/main/app/src/app/dashboard/model/dashboard-card-type.ts
/*
 * SPDX-FileCopyrightText: 2022 Atos
 * SPDX-License-Identifier: EUPL-1.2+
 */

/* These are the available dashboard card implementations.*/
export enum DashboardCardType {
  INFORMATIEOBJECTEN = "INFORMATIEOBJECTEN",
  TAKEN = "TAKEN",
  TAAK_ZOEKEN = "TAAK_ZOEKEN",
  ZAKEN = "ZAKEN",
  ZAAK_WAARSCHUWINGEN = "ZAAK_WAARSCHUWINGEN",
  ZAAK_ZOEKEN = "ZAAK_ZOEKEN",
}