// src/main/app/src/app/dashboard/model/dashboard-card-id.ts
/*
 * SPDX-FileCopyrightText: 2022 Atos
 * SPDX-License-Identifier: EUPL-1.2+
 */

/* These are the available dashboard cards.*/
export enum DashboardCardId {
    MIJN_DOCUMENTEN_NIEUW = 'MIJN_DOCUMENTEN_NIEUW',
    MIJN_TAKEN = 'MIJN_TAKEN',
    MIJN_TAKEN_NIEUW = 'MIJN_TAKEN_NIEUW',
    MIJN_ZAKEN = 'MIJN_ZAKEN',
    MIJN_ZAKEN_NIEUW = 'MIJN_ZAKEN_NIEUW',
    MIJN_ZAKEN_WAARSCHUWING = 'MIJN_ZAKEN_WAARSCHUWING'
}